requirements.txt
tweepy==4.4.0
transformers==4.22.0
pandas==1.3.3
python-dotenv==0.19.0
torch==1.10.0